feat: add mypy plugin for modified Pydantic base model metaclass
With the change in 4.2.0 library clients are able to refer to the OpenEPD models via Class.field_name notation, an extension to standard behaviour of pydantic models. This, however, breaks compatibility with the pydantic mypy plugin.

Expose an improved version of the plugin which understands original Pydantic models and openEPD models.