feat: add exclusive group validator for exposure classes
For concretes, exposure classes are expressed as a list of enum values, for example `csa_exposure_class: list[CsaExposureClass]`. However, some values in a group are mutually exclusive, for example it should not be possible to supply both the `C1` and `C2` values at the same time as they describe resistance to chlorides and freezing.

asana: [openEPD exposure classes advanced validation](https://app.asana.com/0/1203527467850668/1208255682745232/f)